requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.22.0
gunicorn>=21.2.0 
pydantic>=2.0.0
httpx>=0.24.1
throttled>=0.2.1
dependency-injector>=4.46.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langchain-community>=0.0.13
langchain-groq>=0.2.5
langgraph-supervisor>=0.0.9

# For API Calls
requests>=2.31.0

# Visualization dependencies
pillow>=9.0.0
imageio>=2.25.1
numpy>=1.22.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
coverage>=7.6.12
pytest-cov>=6.0.0
pytest-mock>=3.14.0

# Code formatting
black>=25.1.0

# Environment variables
python-dotenv>=1.0.0
